+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3
+
+* Breaking change:
+   *  `getNativeRoots` now returns `StoraxVolume` instead of `Map<String, dynamic>`
+   *  `getAllRoots` now returns `StoraxVolume` instead of `Map<String, dynamic>`
+   *  `listDirectory` now returns `StoraxEntry` instead of `Map<String, dynamic>`
+   *  `traverseDirectory` now returns `StoraxEntry` instead of `Map<String, dynamic>`
+   *  `detectOEM` now returns `StoraxOem` instead of `Map<String, dynamic>`
+
 ## 0.0.2
 
 * Dependency updates.